+++ common/src/main/java/io/github/flemmli97/flan/claim/ClaimStorage.java
@@ -14,6 +14,7 @@ import io.github.flemmli97.flan.api.data.IPlayerData;
 import io.github.flemmli97.flan.api.permission.ClaimPermission;
 import io.github.flemmli97.flan.api.permission.PermissionRegistry;
 import io.github.flemmli97.flan.config.ConfigHandler;
+import io.github.flemmli97.flan.platform.integration.dynmap.DynmapCalls;
 import io.github.flemmli97.flan.platform.integration.permissions.PermissionNodeHandler;
 import io.github.flemmli97.flan.player.EnumDisplayType;
 import io.github.flemmli97.flan.player.EnumEditMode;
@@ -154,6 +155,7 @@ public class ClaimStorage implements IPermissionStorage {
             claim.remove();
             claim.getOwnerPlayer().ifPresent(o -> PlayerClaimData.get(o).updateScoreboard());
         }
+        DynmapCalls.removeMarker(claim);
         return this.claimUUIDMap.remove(claim.getClaimID()) != null;
     }
 
@@ -275,6 +277,7 @@ public class ClaimStorage implements IPermissionStorage {
             old.add(claim);
             return old;
         });
+        DynmapCalls.addClaimMarker(claim);
     }
 
     public boolean transferOwner(Claim claim, ServerPlayer player, UUID newOwner) {
@@ -292,6 +295,7 @@ public class ClaimStorage implements IPermissionStorage {
             return old;
         });
         this.dirty.add(claim.getOwner());
+        DynmapCalls.changeClaimOwner(claim);
         return true;
     }