+++ forge/src/main/java/io/github/flemmli97/flan/forge/FlanForge.java
@@ -6,6 +6,7 @@ import io.github.flemmli97.flan.forge.forgeevent.EntityInteractEventsForge;
 import io.github.flemmli97.flan.forge.forgeevent.ItemInteractEventsForge;
 import io.github.flemmli97.flan.forge.forgeevent.ServerEvents;
 import io.github.flemmli97.flan.forge.forgeevent.WorldEventsForge;
+import io.github.flemmli97.flan.platform.integration.dynmap.DynmapIntegration;
 import io.github.flemmli97.flan.scoreboard.ClaimCriterias;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.EventPriority;
@@ -49,6 +50,9 @@ public class FlanForge {
         forge.addListener(ServerEvents::serverFinishLoad);
         forge.addListener(ServerEvents::disconnect);
         forge.addListener(ServerEvents::serverTick);
+
+        if (ModList.get().isLoaded("dynmap"))
+            DynmapIntegration.reg();
 
         ClaimCriterias.init();
     }